get rid of precompiled files
