+++ src/chunk-streamer.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "precompiled.h"
+#include "main.h"
 
 #include "chunk-streamer.h"
 #include "core.h"
