+++ src/common.h
@@ -173,4 +173,16 @@ namespace boost { namespace geometry { namespace traits {
 	};
 }}}
 
+struct pair_hash
+{
+	template <class T1, class T2>
+	std::size_t operator () (std::pair<T1, T2> const& pair) const
+	{
+		std::size_t h1 = std::hash<T1>()(pair.first);
+		std::size_t h2 = std::hash<T2>()(pair.second);
+
+		return h1 ^ h2;
+	}
+};
+
 #endif
