+++ src/grid.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "precompiled.h"
+#include "main.h"
 
 #include "grid.h"
 #include "core.h"
