+++ src/manipulation/float.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "../precompiled.h"
+#include "../main.h"
 
 #include "float.h"
 #include "../core.h"
