--- src/precompiled.h
@@ -1,77 +1,0 @@
-/*
- * Copyright (C) 2017 Incognito
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef PRECOMPILED_H
-#define PRECOMPILED_H
-
-template<typename T>
-inline void intrusive_ptr_add_ref(T *t)
-{
-	++t->references;
-}
-
-template<typename T>
-inline void intrusive_ptr_release(T *t)
-{
-	if (!(--t->references))
-	{
-		delete t;
-	}
-}
-
-#include <boost/bimap.hpp>
-#include <boost/bimap/multiset_of.hpp>
-#include <boost/bimap/unordered_set_of.hpp>
-#include <boost/geometry.hpp>
-#include <boost/geometry/geometries/geometries.hpp>
-
-#include <Eigen/Core>
-
-#include <algorithm>
-#include <bitset>
-#include <cmath>
-#include <functional>
-#include <limits>
-#include <map>
-#include <queue>
-#include <set>
-#include <sstream>
-#include <string>
-#include <vector>
-#include <utility>
-#include <unordered_set>
-#include <unordered_map>
-#include <memory>
-#include <chrono>
-#include <variant>
-#include <tuple>
-
-#include "common.h"
-#include "sampgdk.h"
-
-struct pair_hash
-{
-	template <class T1, class T2>
-	std::size_t operator () (std::pair<T1, T2> const& pair) const
-	{
-		std::size_t h1 = std::hash<T1>()(pair.first);
-		std::size_t h2 = std::hash<T2>()(pair.second);
-
-		return h1 ^ h2;
-	}
-};
-
-#endif
